apps/phases: move InlineFormSet validation in own class and use in dashboard and admin
